backend/requirements.txt
Flask==2.3.0
flask-cors==4.0.0
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
cryptography==41.0.7

